fixed format

+++ src/Verifier.sol
@@ -22,17 +22,17 @@ pragma solidity >=0.7.0 <0.9.0;
 
 contract Groth16Verifier {
     // Scalar field size
-    uint256 constant r    = 21888242871839275222246405745257275088548364400416034343698204186575808495617;
+    uint256 constant r = 21888242871839275222246405745257275088548364400416034343698204186575808495617;
     // Base field size
-    uint256 constant q   = 21888242871839275222246405745257275088696311157297823662689037894645226208583;
+    uint256 constant q = 21888242871839275222246405745257275088696311157297823662689037894645226208583;
 
     // Verification Key data
-    uint256 constant alphax  = 15678020696299724570210779014841011252140168630041796256896620038769402862137;
-    uint256 constant alphay  = 2285074671128337157625989511519518618993781142221557177793337694814166164410;
-    uint256 constant betax1  = 4958791286074018029923442479591676994406494762394962676102593845160281987543;
-    uint256 constant betax2  = 10875379180077697139234443160697355903104553488303512818324298075157709771414;
-    uint256 constant betay1  = 6416074368858545652461141216030342335690388029903019984422873669725533517321;
-    uint256 constant betay2  = 6598860559337799970306903628821108403705454827779924131757244118591303071680;
+    uint256 constant alphax = 15678020696299724570210779014841011252140168630041796256896620038769402862137;
+    uint256 constant alphay = 2285074671128337157625989511519518618993781142221557177793337694814166164410;
+    uint256 constant betax1 = 4958791286074018029923442479591676994406494762394962676102593845160281987543;
+    uint256 constant betax2 = 10875379180077697139234443160697355903104553488303512818324298075157709771414;
+    uint256 constant betay1 = 6416074368858545652461141216030342335690388029903019984422873669725533517321;
+    uint256 constant betay2 = 6598860559337799970306903628821108403705454827779924131757244118591303071680;
     uint256 constant gammax1 = 11559732032986387107991004021392285783925812861821192530917403151452391805634;
     uint256 constant gammax2 = 10857046999023057135944570762232829481370756359578518086990519993285655852781;
     uint256 constant gammay1 = 4082367875863433681332203403145435568316851327593401208105741076214120093531;
@@ -42,27 +42,30 @@ contract Groth16Verifier {
     uint256 constant deltay1 = 16569813597714576420865324551695550881571845123930970837929585262659715250293;
     uint256 constant deltay2 = 21777002075969881667951684241418148061297974478724808143490770436158537991505;
 
-    
     uint256 constant IC0x = 5385103350887010830873108627668922695843047731102671077192340261102173083864;
     uint256 constant IC0y = 14486705117280963363152081283943228135614805867562991299313642623598526663905;
-    
+
     uint256 constant IC1x = 4234552536175153278668117860769786366623199053653517022048381270698901096009;
     uint256 constant IC1y = 18316225377588811987396526679707772634359316295124170077887796478716738857102;
-    
+
     uint256 constant IC2x = 2319210624017782724820157958295173930721551537594648276525285841703751771638;
     uint256 constant IC2y = 18724795430105287864432808918773415596673864362126981462323337610494894737024;
-    
+
     uint256 constant IC3x = 1829961014563647307894112121777282961632414291612629415299489992810814824655;
     uint256 constant IC3y = 4287125553078873523420528446463309620820793443604810984718314434906285389088;
-    
- 
+
     // Memory data
     uint16 constant pVk = 0;
     uint16 constant pPairing = 128;
 
     uint16 constant pLastMem = 896;
 
-    function verifyProof(uint[2] calldata _pA, uint[2][2] calldata _pB, uint[2] calldata _pC, uint[3] calldata _pubSignals) public view returns (bool) {
+    function verifyProof(
+        uint256[2] calldata _pA,
+        uint256[2][2] calldata _pB,
+        uint256[2] calldata _pC,
+        uint256[3] calldata _pubSignals
+    ) public view returns (bool) {
         assembly {
             function checkField(v) {
                 if iszero(lt(v, r)) {
@@ -70,7 +73,7 @@ contract Groth16Verifier {
                     return(0, 0x20)
                 }
             }
-            
+
             // G1 function to multiply a G1 value(x,y) to value in an address
             function g1_mulAccC(pR, x, y, s) {
                 let success
@@ -105,13 +108,12 @@ contract Groth16Verifier {
                 mstore(add(_pVk, 32), IC0y)
 
                 // Compute the linear combination vk_x
-                
+
                 g1_mulAccC(_pVk, IC1x, IC1y, calldataload(add(pubSignals, 0)))
-                
+
                 g1_mulAccC(_pVk, IC2x, IC2y, calldataload(add(pubSignals, 32)))
-                
+
                 g1_mulAccC(_pVk, IC3x, IC3y, calldataload(add(pubSignals, 64)))
-                
 
                 // -A
                 mstore(_pPairing, calldataload(pA))
@@ -137,7 +139,6 @@ contract Groth16Verifier {
                 mstore(add(_pPairing, 384), mload(add(pMem, pVk)))
                 mstore(add(_pPairing, 416), mload(add(pMem, add(pVk, 32))))
 
-
                 // gamma2
                 mstore(add(_pPairing, 448), gammax1)
                 mstore(add(_pPairing, 480), gammax2)
@@ -154,7 +155,6 @@ contract Groth16Verifier {
                 mstore(add(_pPairing, 704), deltay1)
                 mstore(add(_pPairing, 736), deltay2)
 
-
                 let success := staticcall(sub(gas(), 2000), 8, _pPairing, 768, _pPairing, 0x20)
 
                 isOk := and(success, mload(_pPairing))
@@ -164,21 +164,20 @@ contract Groth16Verifier {
             mstore(0x40, add(pMem, pLastMem))
 
             // Validate that all evaluations ∈ F
-            
+
             checkField(calldataload(add(_pubSignals, 0)))
-            
+
             checkField(calldataload(add(_pubSignals, 32)))
-            
+
             checkField(calldataload(add(_pubSignals, 64)))
-            
+
             checkField(calldataload(add(_pubSignals, 96)))
-            
 
             // Validate all evaluations
             let isValid := checkPairing(_pA, _pB, _pC, _pubSignals, pMem)
 
             mstore(0, isValid)
-             return(0, 0x20)
-         }
-     }
- }
+            return(0, 0x20)
+        }
+    }
+}
